Extract HTML generation into a separate renderer
