+++ app/templates/output_document_introduction.html.erb
@@ -5,16 +5,16 @@
     <div class="appointment-details">
       <p class="appointment-details__label">Date of your appointment</p>
 
-      <p class="appointment-details__data appointment-details__data--date"><%= appointment_date %></p>
+      <p class="appointment-details__data appointment-details__data--date"><%= output_document.appointment_date %></p>
 
       <p class="appointment-details__label">Your guidance specialist was</p>
 
-      <p class="appointment-details__data"><%= guider_first_name %></p>
+      <p class="appointment-details__data"><%= output_document.guider_first_name %></p>
 
       <p class="appointment-details__label">from</p>
 
       <p class="appointment-details__data">
-        <%= guider_organisation %>
+        <%= output_document.guider_organisation %>
       </p>
     </div>
   </div>
@@ -22,7 +22,7 @@
   <div class="l-record-info">
     <div class="record-info">
       <p class="record-info__name">
-        <%= attendee_name %>
+        <%= output_document.attendee_name %>
       </p>
       <p class="record-info__title">
         Record of your Pension Wise appointment
@@ -49,7 +49,7 @@
       <p class="total-value__label">Total value of your pension pot(s)</p>
 
       <p class="total-value__amount">
-        <%= value_of_pension_pots %>
+        <%= output_document.value_of_pension_pots %>
       </p>
 
       <p class="total-value__note">
@@ -64,12 +64,12 @@
   <div class="l-grid-row">
     <div class="l-column-1">
 
-      <% if income_in_retirement == 'pension' %>
+      <% if output_document.income_in_retirement == 'pension' %>
         <p>
           You told us that your defined contribution pot is your only income in
           retirement.
         </p>
-      <% elsif income_in_retirement == 'other' %>
+      <% elsif output_document.income_in_retirement == 'other' %>
         <p>
           You told us that you have other income in retirement, eg from a
           defined benefit pension
@@ -103,7 +103,7 @@
       <div class="l-text-narrow emphasize">
         <h3>Things to talk about</h3>
 
-        <% if income_in_retirement == 'pension' %>
+        <% if output_document.income_in_retirement == 'pension' %>
           <p>How much tax you’ll pay - the tax-free Personal Allowance for the
             2015/16 tax year is £10,600. </p>
 
@@ -122,7 +122,7 @@
             might be eligible for and how to claim.</p>
 
 
-        <% elsif income_in_retirement == 'other' %>
+        <% elsif output_document.income_in_retirement == 'other' %>
           <p>Tax you pay when taking additional income from your pension.</p>
 
           <p>How long your money needs to last - on average, people aged 55
